fnc_matrizOD/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
json
geojson
haversine
osmnx
geopandas
shapely.geometry
pandas
networkx
requests
folium
BytesIO